Moving a filter to a new tag file to fix problem with it loading
