+++ portal/plugins/gnmsyndication/templatetags/syndicationstats_customfilters.py
@@ -49,30 +49,6 @@ def syndicationStatusFormatter(value):
 
     return mark_safe(u"<span class=\"{0}\">{1}&nbsp;&nbsp;{2}</span>".format(textforpassing,iconstr,value))
 
-@register.filter("platformindicator")
-def platformIndicator(value):
-    #if isinstance(value,list):
-    #    return " | ".join(value)
-    #else:
-    #    return value
-
-    logging.debug("platformindicator: upload intentions: {0}".format(value))
-    if not isinstance(value,list):
-        value = [value]
-
-    lowered_values = []
-    for v in value:
-        lowered_values.append(v.lower())
-
-    rtn = u""
-
-    for p in models.platform.objects.all():
-        if p.intention_label in lowered_values:
-            rtn += u"<img class=\"tooltip_icon inline_medium_icon\" src=\"{0}\" alt=\"{1} enabled\">".format(p.enabled_icon_url,p.name)
-        else:
-            rtn += u"<img class=\"tooltip_icon inline_medium_icon\" src=\"{0}\" alt=\"{1} disabled\">".format(p.disable_icon_url,p.name)
-
-    return mark_safe(rtn)
 
 @register.filter("pacformindicator")
 def pacformIndicator(value):
